server code cleanup
